Split core and filesystem into crates
