+++ crates/node-bindings/src/core/js_requests/entry_request/mod.rs
@@ -3,8 +3,8 @@ use std::rc::Rc;
 use napi::Env;
 use napi::JsObject;
 use napi_derive::napi;
+use parcel_napi_helpers::anyhow_napi;
 
-use crate::core::js_helpers::anyhow_napi;
 use crate::core::js_requests::request_options::input_fs_from_options;
 use crate::core::requests::entry_request::run_entry_request;
 use crate::core::requests::entry_request::EntryRequestInput;
