+++ crates/node-bindings/src/core/mod.rs
@@ -1,7 +1,5 @@
 //! Core re-implementation in Rust
 
-mod filesystem;
-mod js_helpers;
 mod js_requests;
 mod project_path;
 /// Request types and run functions
