+++ crates/parcel_core/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "parcel_core"
+version = "0.1.0"
+edition = "2021"
+description = "Parcel bundler core logic"
+
+[dependencies]
