+++ crates/parcel_filesystem/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "parcel_filesystem"
+version = "0.1.0"
+edition = "2021"
+description = "FileSystem wrapper trait for use in Parcel codebase."
+
+[dependencies]
+parcel_napi_helpers = { path = "../parcel_napi_helpers" }
+napi = "2.16.4"
+dashmap = "5.5.3"
+anyhow = "1.0.82"
+
+[dev-dependencies]
+assert_fs = "1.1.1"
