+++ crates/parcel_napi_helpers/src/lib.rs
@@ -7,7 +7,7 @@ use napi::NapiRaw;
 
 /// Convert anyhow error to napi error
 pub fn anyhow_napi(value: anyhow::Error) -> napi::Error {
-  napi::Error::from_reason(format!("[napi] {}", value.to_string()))
+    napi::Error::from_reason(format!("[napi] {}", value.to_string()))
 }
 
 /// Get an object field as a JSFunction. Will error out if the field is not present or isn't an
@@ -16,23 +16,23 @@ pub fn anyhow_napi(value: anyhow::Error) -> napi::Error {
 /// ## Safety
 /// Uses raw NAPI casts, but checks that object field is a function
 pub fn get_function(env: &Env, js_object: &JsObject, field_name: &str) -> napi::Result<JsFunction> {
-  let Some(method): Option<JsUnknown> = js_object.get(field_name)? else {
-    return Err(napi::Error::from_reason(format!(
-      "[napi] Method not found: {}",
-      field_name
-    )));
-  };
-  let function_class: JsUnknown = env.get_global()?.get_named_property("Function")?;
-  let is_function = method.instanceof(function_class)?;
-  if !is_function {
-    return Err(napi::Error::from_reason(format!(
-      "[napi] Method is not a function: {}",
-      field_name
-    )));
-  }
+    let Some(method): Option<JsUnknown> = js_object.get(field_name)? else {
+        return Err(napi::Error::from_reason(format!(
+            "[napi] Method not found: {}",
+            field_name
+        )));
+    };
+    let function_class: JsUnknown = env.get_global()?.get_named_property("Function")?;
+    let is_function = method.instanceof(function_class)?;
+    if !is_function {
+        return Err(napi::Error::from_reason(format!(
+            "[napi] Method is not a function: {}",
+            field_name
+        )));
+    }
 
-  let method_fn = unsafe { JsFunction::from_napi_value(env.raw(), method.raw()) }?;
-  Ok(method_fn)
+    let method_fn = unsafe { JsFunction::from_napi_value(env.raw(), method.raw()) }?;
+    Ok(method_fn)
 }
 
 /// Call a method on an object with a set of arguments.
@@ -57,12 +57,12 @@ pub fn get_function(env: &Env, js_object: &JsObject, field_name: &str) -> napi::
 /// call_method(&self.env, &js_object, field_name, &args)?;
 /// ```
 pub fn call_method(
-  env: &Env,
-  js_object: &JsObject,
-  field_name: &str,
-  args: &[&JsUnknown],
+    env: &Env,
+    js_object: &JsObject,
+    field_name: &str,
+    args: &[&JsUnknown],
 ) -> napi::Result<JsUnknown> {
-  let method_fn = get_function(env, js_object, field_name)?;
-  let result = method_fn.call(Some(&js_object), &args)?;
-  Ok(result)
+    let method_fn = get_function(env, js_object, field_name)?;
+    let result = method_fn.call(Some(&js_object), &args)?;
+    Ok(result)
 }
